requirements.txt
transformers
torch
openai
mysql-connector-python
flask
flask-Cors
urllib3